+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 ### Fixed
+- Plugin compatibility with Plover v4.0.0.dev11
+
 ### Removed
 
 ## [0.1.0] - 2022-01-02